Refactor function signatures to use int64_t for size and index parameters
- Updated the return type of `font` function to int and adjusted error handling.
- Changed `strlen` to return int64_t instead of int.
- Modified `instr`, `strmid`, and `getstr` functions to accept and return int64_t.
- Updated `peek`, `poke`, `memcpy`, and `memset` implementations to use int64_t for index and size parameters.
- Adjusted related function calls in the DemoDrawExtended and ApiCompileTest files to accommodate the new int64_t type.

+++ HspppLib/src/core/hsppp_math.inl
@@ -113,8 +113,8 @@ namespace hsppp {
         return std::to_string(value);
     }
 
-    int strlen(const std::string& p1) {
-        return static_cast<int>(p1.size());
+    int64_t strlen(const std::string& p1) {
+        return static_cast<int64_t>(p1.size());
     }
 
     // ============================================================
